Types

Index

EnergyModelsBase.AbstractStorageParameters

EnergyModelsBase.Formulation

EnergyModelsBase.Node

EnergyModelsBase.StorageBehavior

Nodal and link supertypes

EnergyModelsBase.Node — Type

Node as supertype for all technologies.

EnergyModelsBase.Formulation — Type

Declaration of the general type for formulation of links.

Parameter supertypes

EnergyModelsBase.StorageBehavior — Type

abstract type StorageBehavior

StorageBehavior as supertype for individual storage behaviours.

Storage behaviour is used to identify how a storage node should behave within the individual TimeStructures of a strategic period.

EnergyModelsBase.AbstractStorageParameters — Type

AbstractStorageParameters as supertype for individual parameters for Storage nodes.

Storage parameters are used to provide the user the flexibility to include or not include capacities and variable and fixed OPEX parameters for charging, the storage level, and discharging.