Resources

Resources correspond to the mass/energy that is converted or transported within an energy system. Resources are discrete, that is they do not have as default additional variables, e.g. pressure or temperature, associated with them. Instead, they are implemented through flows and levels, as explained in Optimization variables.

Index

EnergyModelsBase.Resource

EnergyModelsBase.ResourceCarrier

EnergyModelsBase.ResourceEmit

EnergyModelsBase.co2_int

Resource types

The following resources are implemented in EnergyModelsBase. EnergyModelsBase differentiates between ResourceCarrier and ResourceEmit resources. The key difference between both is that ResourceEmit resources can have emissions, e.g., CO₂ or methane. Emissions are accounted for and can have either a cap and/or a price associated with them.

One important field for a resource is the CO₂ intensity (co2_int). CO₂ is handled differently than other emissions as the emissions are fundamental properties of a fuel based on the carbon content.

EnergyModelsBase.Resource — Type

General resource supertype to be used for the declaration of subtypes.

EnergyModelsBase.ResourceCarrier — Type

ResourceCarrier{T<:Real} <: Resource

Resources that can be transported and converted. These resources cannot be included as resources that are emitted, e.g, in the variable emissions_strategic.

Fields

id is the name/identifyer of the resource.

co2_int::T is the CO₂ intensity, e.g., t/MWh.

EnergyModelsBase.ResourceEmit — Type

ResourceEmit{T<:Real} <: Resource

Resources that can be emitted (e.g., CO₂, CH₄, NOₓ).

These resources can be included as resources that are emitted, e.g, in the variable emissions_strategic.

Fields

id is the name/identifyer of the resource.

co2_int::T is the CO₂ intensity, e.g., t/MWh.

Functions for accessing fields of Resource types

The following functions are declared for accessing fields from a Resource type. If you want to introduce new Resource types, it is important that this function are either functional for your new types or you have to declare a corresponding function.

EnergyModelsBase.co2_int — Function

co2_int(p::Resource)

Returns the CO₂ intensity of resource p