Internals - EnergyModelsInvestment extension

Index

EMIExt.InvestmentModel

EMIExt.SingleInvData

EMIExt.StorageInvData

EnergyModelsInvestments.BinaryInvestment

EnergyModelsInvestments.ContinuousInvestment

EnergyModelsInvestments.DiscreteInvestment

EnergyModelsInvestments.FixedInvestment

EnergyModelsInvestments.PeriodLife

EnergyModelsInvestments.RollingLife

EnergyModelsInvestments.SemiContinuousInvestment

EnergyModelsInvestments.SemiContinuousOffsetInvestment

EnergyModelsInvestments.StudyLife

EMIExt.check_inv_data

EnergyModelsBase.check_node_data

EnergyModelsBase.constraints_capacity_installed

EnergyModelsBase.objective

EnergyModelsBase.variables_capex

EnergyModelsInvestments.has_investment

EnergyModelsInvestments.investment_data

Extension

Types

EMIExt.InvestmentModel — Type

InvestmentModel <: EMB.InvestmentModel

Internal type for InvestmentModel. The introduction of an internal type is necessary as extensions do not allow to export functions or types.

EMIExt.SingleInvData — Type

SingleInvData <: EMB.SingleInvData

Internal type for SingleInvData. The introduction of an internal type is necessary as extensions do not allow to export functions or types.

EMIExt.StorageInvData — Type

StorageInvData <: EMB.StorageInvData

Internal type for StorageInvData. The introduction of an internal type is necessary as extensions do not allow to export functions or types.

Methods

EMIExt.check_inv_data — Function

check_inv_data(
    inv_data::AbstractInvData,
    capacity_profile::TimeProfile,
    𝒯,
    message::String,
    check_timeprofiles::Bool,
)

Performs various checks on investment data introduced within EnergyModelsInvestments

Checks

For each field with TimeProfile:

If the TimeProfile is a StrategicProfile, it will check that the profile is in accordance with the TimeStructure

TimeProfiles in InvestmentData cannot include OperationalProfile, RepresentativeProfile, or ScenarioProfile as this is not allowed through indexing on the TimeProfile.

The field :min_add has to be less than :max_add if the investment mode is given by ContinuousInvestment or SemiContiInvestment.

Existing capacity cannot be larger than :max_inst capacity in the beginning. If NoStartInvData is used, it also checks that the the TimeProfile capacity_profile is not including OperationalProfile, RepresentativeProfile, or ScenarioProfile to avoid indexing problems.

EnergyModelsBase

Methods

EnergyModelsBase.variables_capex — Method

EMB.variables_capex(m, 𝒩, 𝒯, 𝒫, modeltype::AbstractInvestmentModel)

Create variables for the capital costs for the investments in storage and technology nodes.

Additional variables for investment in capacity:

:cap_capex - CAPEX costs for a technology

:cap_current - installed capacity for storage in each strategic period

:cap_add - added capacity

:cap_rem - removed capacity

:cap_invest_b - binary variable whether investments in capacity are happening

:cap_remove_b - binary variable whether investments in capacity are removed

Additional variables for investment in storage:

:stor_level_capex - CAPEX costs for increases in the capacity of a storage

:stor_level_current - installed capacity for storage in each strategic period

:stor_level_add - added capacity

:stor_level_rem - removed capacity

:stor_level_invest_b - binary variable whether investments in capacity are happening

:stor_level_remove_b - binary variable whether investments in capacity are removed

:stor_charge_capex - CAPEX costs for increases in the rate of a storage

:stor_charge_current - installed rate for storage in each strategic period

:stor_charge_add - added rate

:stor_charge_rem - removed rate

:stor_charge_invest_b - binary variable whether investments in rate are happening

:stor_charge_remove_b - binary variable whether investments in rate are removed

EnergyModelsBase.constraints_capacity_installed — Method

constraints_capacity_installed(m, n::Node, 𝒯::TimeStructure, modeltype::AbstractInvestmentModel)
constraints_capacity_installed(m, n::Storage, 𝒯::TimeStructure, modeltype::AbstractInvestmentModel)

When the modeltype is an investment model, the function introduces the related constraints for the capacity expansion. The investment mode and lifetime mode are used for adding constraints.

The default function only accepts nodes with SingleInvData. If you have several capacities for investments, you have to dispatch specifically on the node type. This is implemented for Storage nodes where the function introduces the related constraints for the capacity expansions for the fields :charge, :level, and :discharge. This requires the utilization of the StorageInvData investment type, in which the investment mode and lifetime mode are used for adding constraints for each capacity.

EnergyModelsBase.objective — Method

EMB.objective(m, 𝒩, 𝒯, 𝒫, modeltype::AbstractInvestmentModel)

Create objective function overloading the default from EMB for AbstractInvestmentModel.

Maximize Net Present Value from investments (CAPEX) and operations (OPEX and emission costs)

TODO:

Consider adding contributions from

revenue (as positive variable, adding positive)

maintenance based on usage (as positive variable, adding negative)

These variables would need to be introduced through the package SparsVariables.

Both are not necessary, as it is possible to include them through the OPEX values, but it would be beneficial for a better separation and simpler calculations from the results.

EnergyModelsBase.check_node_data — Method

EMB.check_node_data(n::EMB.Node, data::InvestmentData, 𝒯, modeltype::AbstractInvestmentModel, check_timeprofiles::Bool)
EMB.check_node_data(n::Storage, data::InvestmentData, 𝒯, modeltype::AbstractInvestmentModel, check_timeprofiles::Bool)

Performs various checks on investment data for standard and Storage nodes.

Checks for standard nodes

Each node can only have a single InvestmentData.

All checks incorporated in the function check_inv_data.

Checks for Storage nodes

Each node can only have a single InvestmentData.

The InvestmentData must be StorageInvData.

For each individual investment field all checks incorporated in the function check_inv_data.

EnergyModelsInvestments

Constructors

The following constructors are only relevant for the legacy constructors introduced within the extension. They do not provide any additional information.

EnergyModelsInvestments.BinaryInvestment — Type

When the field cap is not included, it is assumed that its value is FixedProfile(0). This behavior is only for allowing the legacy constructor to work, while it will be removed in the near future.

EnergyModelsInvestments.ContinuousInvestment — Type

When the fields min_add and max_add are not included, it is assumed that their values are FixedProfile(0). This behavior is only for allowing the legacy constructor to work, while it will be removed in the near future.

EnergyModelsInvestments.DiscreteInvestment — Type

When the field increment is not included, it is assumed that its value is FixedProfile(0). This behavior is only for allowing the legacy constructor to work, while it will be removed in the near future.

EnergyModelsInvestments.FixedInvestment — Type

When the field cap is not included, it is assumed that its value is FixedProfile(0). This behavior is only for allowing the legacy constructor to work, while it will be removed in the near future.

EnergyModelsInvestments.PeriodLife — Type

When the field lifetime is not included, it is assumed that its value is FixedProfile(0). This behavior is only for allowing the legacy constructor to work, while it will be removed in the near future.

EnergyModelsInvestments.RollingLife — Type

When the field lifetime is not included, it is assumed that its value is FixedProfile(0). This behavior is only for allowing the legacy constructor to work, while it will be removed in the near future.

EnergyModelsInvestments.SemiContinuousInvestment — Type

When the fields min_add and max_add are not included, it is assumed that their values are FixedProfile(0). This behavior is only for allowing the legacy constructor to work, while it will be removed in the near future.

EnergyModelsInvestments.SemiContinuousOffsetInvestment — Type

When the fields min_add, max_add, and capex_offset are not included, it is assumed that their values are FixedProfile(0). This behavior is only for allowing the legacy constructor to work, while it will be removed in the near future.

EnergyModelsInvestments.StudyLife — Type

When the field lifetime is not included, it is assumed that its value is FixedProfile(0). This behavior is only for allowing the legacy constructor to work, while it will be removed in the near future.

Methods

EnergyModelsInvestments.has_investment — Function

EMI.has_investment(n::EMB.Node)

For a given Node n, checks that it contains the required investment data.

EMI.has_investment(n::Storage, field::Symbol)

For a given Storage node, checks that it contains investments for the field field, that is :charge, :level, or :discharge.

EnergyModelsInvestments.investment_data — Function

EMI.investment_data(inv_data::SingleInvData)

Return the investment data of the investment data SingleInvData.

EMI.investment_data(n::EMB.Node)
EMI.investment_data(n::EMB.Node, field::Symbol)

Return the InvestmentData of the Node n or if field is specified, it returns the InvData for the corresponding capacity.